lazy load the terriajs so loader shows immediately
