+++ lib/Views/UserInterface.jsx
@@ -1,13 +1,16 @@
 import PropTypes from "prop-types";
+import React, { Suspense } from "react";
 import { useSyncExternalStore } from "react";
-import RelatedMaps from "terriajs/lib/ReactViews/RelatedMaps/RelatedMaps";
-import { MenuLeft } from "terriajs/lib/ReactViews/StandardUserInterface/customizable/Groups";
-import MenuItem from "terriajs/lib/ReactViews/StandardUserInterface/customizable/MenuItem";
-import StandardUserInterface from "terriajs/lib/ReactViews/StandardUserInterface/StandardUserInterface";
-import version from "../../version";
 import "./global.scss";
 import { Loader } from "./Loader";
 import { terriaStore } from "./terriaStore";
+
+// Lazy load the entire TerriaUserInterface component
+const LazyTerriaUserInterface = React.lazy(() =>
+  import("./TerriaUserInterface").then((module) => ({
+    default: module.TerriaUserInterface
+  }))
+);
 
 export const UserInterface = ({ themeOverrides }) => {
   const { terria, viewState, status } = useSyncExternalStore(
@@ -18,31 +21,15 @@ export const UserInterface = ({ themeOverrides }) => {
   if (status === "loading") {
     return <Loader />;
   }
-  const relatedMaps = viewState.terria.configParameters.relatedMaps;
-  const aboutButtonHrefUrl =
-    viewState.terria.configParameters.aboutButtonHrefUrl;
 
   return (
-    <StandardUserInterface
-      terria={terria}
-      viewState={viewState}
-      themeOverrides={themeOverrides}
-      version={version}
-    >
-      <MenuLeft>
-        {aboutButtonHrefUrl ? (
-          <MenuItem
-            caption="About"
-            href={aboutButtonHrefUrl}
-            key="about-link"
-          />
-        ) : null}
-        {relatedMaps && relatedMaps.length > 0 ? (
-          <RelatedMaps relatedMaps={relatedMaps} />
-        ) : null}
-      </MenuLeft>
-      {/* <ExperimentalMenu /> */}
-    </StandardUserInterface>
+    <Suspense fallback={<Loader />}>
+      <LazyTerriaUserInterface
+        terria={terria}
+        viewState={viewState}
+        themeOverrides={themeOverrides}
+      />
+    </Suspense>
   );
 };
